chore: another small cleanup Signed-off-by: benjdlambert <ben@blam.sh>
