+++ plugins/auth-backend/src/service/OidcRouter.ts
@@ -1,0 +1,73 @@
+/*
+ * Copyright 2025 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import Router from 'express-promise-router';
+import { OidcService } from './OidcService';
+import { AuthenticationError } from '@backstage/errors';
+import { AuthService } from '@backstage/backend-plugin-api';
+import { TokenIssuer } from '../identity/types';
+import { UserInfoDatabase } from '../database/UserInfoDatabase';
+
+export class OidcRouter {
+  private constructor(private readonly oidc: OidcService) {}
+
+  static create(options: {
+    auth: AuthService;
+    tokenIssuer: TokenIssuer;
+    baseUrl: string;
+    userInfo: UserInfoDatabase;
+  }) {
+    return new OidcRouter(OidcService.create(options));
+  }
+
+  public getRouter() {
+    const router = Router();
+
+    router.get('/.well-known/openid-configuration', (_req, res) => {
+      res.json(this.oidc.getConfiguration());
+    });
+
+    router.get('/.well-known/jwks.json', async (_req, res) => {
+      const { keys } = await this.oidc.listPublicKeys();
+      res.json({ keys });
+    });
+
+    router.get('/v1/token', (_req, res) => {
+      res.status(501).send('Not Implemented');
+    });
+
+    // This endpoint doesn't use the regular HttpAuthoidc, since the contract
+    // is specifically for the header to be communicated in the Authorization
+    // header, regardless of token type
+    router.get('/v1/userinfo', async (req, res) => {
+      const matches = req.headers.authorization?.match(/^Bearer[ ]+(\S+)$/i);
+      const token = matches?.[1];
+      if (!token) {
+        throw new AuthenticationError('No token provided');
+      }
+
+      const userInfo = await this.oidc.getUserInfo({ token });
+
+      if (!userInfo) {
+        res.status(404).send('User info not found');
+        return;
+      }
+
+      res.json(userInfo);
+    });
+
+    return router;
+  }
+}
